requirements.txt
requests
beautifulsoup4
pandas
spacy
tqdm
python-dotenv
flask